utchs/requirements.txt
numpy>=1.20.0,<2.0.0
matplotlib>=3.5.0,<4.0.0
networkx>=2.7.0,<3.0.0
scipy>=1.8.0,<2.0.0
pyyaml>=6.0,<7.0.0
h5py>=3.6.0,<4.0.0
pillow>=9.0.0,<10.0.0
ipython>=8.0.0,<9.0.0
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
mypy>=1.0.0,<2.0.0
black>=23.0.0,<24.0.0
isort>=5.0.0,<6.0.0
pydantic>=2.0.0,<3.0.0
sphinx>=7.0.0,<8.0.0
sphinx-rtd-theme>=1.3.0,<2.0.0
sphinx-autodoc-typehints>=1.24.0,<2.0.0
sphinx-napoleon>=0.7.0,<1.0.0
